Finished Team view and refactored most of the app
The team view allows a user to create a team, edit
the team's name, and delete the team. It also allows
them to add members (pokemon), personalize them by adding
moves, nicknames, and items. Refactoring converting a lot of
IIFEs and function "components" into class-based components.
Most of these have an update() and a getComponent() method.
They rely on the global state classes for information.
Fixed the login function to ensure it enforces the condition
that a user who's account isn't active isn't able to login.
More refactoring to come.