Refactored usage of Alert class and image caching
Instead of an Alert class from which several alert components
can be made, a global alert class is used, which by default
clears any previous alert before displaying a new one.
Also, instead of adding a time parameter when fetching
a profile image on the client side (to trick browser cache),
the server side appends a Date.now() to filename to avoid
the issue with caching.