Updated API docs, repo link, and updated config
Home page has a link to the website's Github repo. Re-ran docgen
to create a new html file for the backend API docs, which now includes
the /auth/verify/:token route. Lastly, the webpack config was updated
to run in production mode in order to minify the bundled code. The
source map was also commented out to stop mapping errors in production.